backend/pyproject.toml
[tool.poetry]
name = "gdgoc-hackathon-backend"
version = "0.1.0"
description = "Django REST Framework authentication system"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
google-auth = "^2.37"
drf-spectacular = "^0.28"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"
django-cors-headers = "^4.6"
requests = "^2.32.5"
uuid6 = "^2025.0.1"
gunicorn = "^23.0"
django-filter = "^25.2"
google-cloud-storage = "^2.19"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
